backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
channels==4.1.0
channels-redis==4.2.0
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.8
django-celery-beat==2.7.0
vnstock3
daphne==4.1.2
python-decouple==3.8
Pillow==10.4.0
gunicorn==22.0.0
